requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pyautogui>=0.9.54
torch>=2.0.0
torchvision>=0.15.0
